+++ labellerTools.py
@@ -8,23 +8,26 @@ from matplotlib.backend_tools import ToolBase, ToolToggleBase
 from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
 
 class Annotate(object):
-	def __init__(self, canvas, data):
+	def __init__(self, axes, data):
 		self.labeling = False
 		self.data = data
-		self.ax = canvas
+		self.axes = axes
 		self.rect = matplotlib.patches.Rectangle((0,0), 1, 1, color='white', alpha=0.5)
 		self.x0 = None
 		self.y0 = None
 		self.x1 = None
 		self.y1 = None
 		self.pressed = False
-		self.ax.add_patch(self.rect)
-		self.ax.figure.canvas.mpl_connect('button_press_event', self.on_press)
-		self.ax.figure.canvas.mpl_connect('button_release_event', self.on_release)
-		self.ax.figure.canvas.mpl_connect('motion_notify_event', self.on_motion)
+		self.axes.add_patch(self.rect)
+		self.axes.figure.canvas.mpl_connect('button_press_event', self.on_press)
+		self.axes.figure.canvas.mpl_connect('button_release_event', self.on_release)
+		self.axes.figure.canvas.mpl_connect('motion_notify_event', self.on_motion)
 
 	def toggle(self):
 		self.labeling = not self.labeling
+		if not self.labeling:
+			print('toolbar')
+			self.data.saveRect()
 
 	def on_press(self, event):
 		if self.labeling:
@@ -39,8 +42,9 @@ class Annotate(object):
 			self.y1 = event.ydata
 			self.rect.set_width(self.x1 - self.x0)
 			self.rect.set_height(self.y1 - self.y0)
-			self.rect.set_xy((self.x0, self.y0))
-			self.ax.figure.canvas.draw()
+			# get rectangle attributes vertices, coordinates
+			self.data.saveRect(self.rect.get_xy(), self.rect.get_width(), self.rect.get_height())
+			self.axes.figure.canvas.draw()
 
 	def on_motion(self, event):
 		if self.pressed and self.labeling:
@@ -50,7 +54,7 @@ class Annotate(object):
 			self.rect.set_height(self.y1 - self.y0)
 			self.rect.set_xy((self.x0, self.y0))
 			self.rect.set_linestyle('dashed')
-			self.ax.figure.canvas.draw()
+			self.axes.figure.canvas.draw()
 
 
 class MyToolbar(NavigationToolbar):
